+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Bbody.hs
@@ -36,7 +36,6 @@ import Cardano.Ledger.Alonzo.Rules.Utxo (AlonzoUtxoPredFailure)
 import Cardano.Ledger.Alonzo.Rules.Utxos (AlonzoUtxosPredFailure)
 import Cardano.Ledger.Alonzo.Rules.Utxow (AlonzoUtxowPredFailure)
 import Cardano.Ledger.Alonzo.Scripts (ExUnits (..), OrdExUnits (..), pointWiseExUnits)
-import Cardano.Ledger.Alonzo.Tx (totExUnits)
 import Cardano.Ledger.BaseTypes (Mismatch (..), Relation (..), ShelleyBase)
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
@@ -165,7 +164,7 @@ validateExUnits ::
   ExUnits ->
   Rule (EraRule "BBODY" era) 'Transition ()
 validateExUnits txs ppMax =
-  let txTotal = foldMap totExUnits txs
+  let txTotal = foldMap getTotalExUnits txs
    in pointWiseExUnits (<=) txTotal ppMax
         ?! injectFailure
           ( TooManyExUnits $