+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Imp/SubUtxowSpec.hs
@@ -11,9 +11,10 @@ module Test.Cardano.Ledger.Dijkstra.Imp.SubUtxowSpec (spec) where
 
 import Cardano.Ledger.Address (bootstrapKeyHash)
 import Cardano.Ledger.Allegra.Scripts (AllegraEraScript (..))
+import Cardano.Ledger.Alonzo.Plutus.Context (CollectError (..))
 import Cardano.Ledger.Alonzo.Scripts (eraLanguages)
 import Cardano.Ledger.Alonzo.TxWits (unRedeemersL, unTxDatsL)
-import Cardano.Ledger.BaseTypes (Mismatch (..), SlotNo (..), StrictMaybe (..))
+import Cardano.Ledger.BaseTypes (Inject (..), Mismatch (..), SlotNo (..), StrictMaybe (..))
 import Cardano.Ledger.Conway.Governance (
   GovAction (..),
   GovActionId,
@@ -22,10 +23,12 @@ import Cardano.Ledger.Conway.Governance (
   VotingProcedure (..),
   VotingProcedures (..),
  )
+import qualified Cardano.Ledger.Conway.Rules as Conway
 import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential (..), StakeReference (..), credKeyHashWitness)
 import Cardano.Ledger.Dijkstra.Core
 import Cardano.Ledger.Dijkstra.Rules (DijkstraSubUtxowPredFailure (..))
+import Cardano.Ledger.Dijkstra.TxInfo (DijkstraContextError (..))
 import Cardano.Ledger.Keys (asWitness, witVKeyHash)
 import Cardano.Ledger.Plutus (
   Data (..),
@@ -260,9 +263,10 @@ spec = describe "SUBUTXOW" $ do
                        )
             txInAt 0
               <$> withPostFixup (rederiveAddrTxWits . resetTxOutDataHash) (submitTx tx)
-          submitFailingTx
+          submitFailingLegacySubTx
+            lang
             (mkTopTxWithSubTxs [mkBasicTx $ mkBasicTxBody & inputsTxBodyL .~ [txIn]])
-            [injectFailure . SubUnspendableUTxONoDatumHash @era $ NES.singleton txIn]
+            (SubUnspendableUTxONoDatumHash @era $ NES.singleton txIn)
 
   forM_ (eraLanguages @era) $ \lang ->
     withSLanguage lang $ \slang ->
@@ -275,9 +279,10 @@ spec = describe "SUBUTXOW" $ do
           txIn <- produceScript redeemerSameAsDatumHash
           let missingDatum = hashData @era (Data (P.I 3))
           withPostFixupSubTxs (fixupResetAddrWits . (witsTxL . datsTxWitsL .~ mempty)) $
-            submitFailingTx
+            submitFailingLegacySubTx
+              lang
               (mkTopTxWithSubTxs [scriptSpendingSubTx txIn])
-              [injectFailure $ SubMissingRequiredDatums @era (NES.singleton missingDatum) []]
+              (SubMissingRequiredDatums @era (NES.singleton missingDatum) [])
 
         it "SubNotAllowedSupplementalDatums" $ do
           txIn <- produceScript redeemerSameAsDatumHash
@@ -289,11 +294,10 @@ spec = describe "SUBUTXOW" $ do
                         %~ Map.insert extraDatumHash extraDatum
                     )
           withPostFixupSubTxs addExtraDatum $
-            submitFailingTx
+            submitFailingLegacySubTx
+              lang
               (mkTopTxWithSubTxs [scriptSpendingSubTx txIn])
-              [ injectFailure $
-                  SubNotAllowedSupplementalDatums @era (NES.singleton extraDatumHash) []
-              ]
+              (SubNotAllowedSupplementalDatums @era (NES.singleton extraDatumHash) [])
 
         it "SubMissingRedeemers" $ do
           txIn <- produceScript redeemerSameAsDatumHash
@@ -301,7 +305,8 @@ spec = describe "SUBUTXOW" $ do
           withPostFixupSubTxs (fixupResetAddrWits . (witsTxL . rdmrsTxWitsL .~ mempty)) $
             submitFailingTx
               (mkTopTxWithSubTxs [scriptSpendingSubTx txIn])
-              [ injectFailure $
+              [ injectFailure $ Conway.CollectErrors [NoRedeemer missingRedeemer]
+              , injectFailure $
                   SubMissingRedeemers @era [(missingRedeemer, redeemerSameAsDatumHash)]
               ]
 
@@ -315,9 +320,10 @@ spec = describe "SUBUTXOW" $ do
                         %~ Map.insert extraPurpose (redeemerData, ExUnits 0 0)
                     )
           withPostFixupSubTxs addExtraRedeemer $
-            submitFailingTx
+            submitFailingLegacySubTx
+              lang
               (mkTopTxWithSubTxs [scriptSpendingSubTx txIn])
-              [injectFailure $ SubExtraRedeemers @era [extraPurpose]]
+              (SubExtraRedeemers @era [extraPurpose])
 
         describe "SubScriptIntegrityHashMismatch" $ do
           let testHashMismatch badHash = do
@@ -337,22 +343,20 @@ spec = describe "SUBUTXOW" $ do
                   rederiveAddrTxWits $
                     fixedUpTx & bodyTxL . subTransactionsTxBodyL .~ OMap.singleton badSubTx
                 withNoFixup $
-                  submitFailingTx
-                    badTopTx
-                    [ injectFailure $
-                        SubScriptIntegrityHashMismatch @era
-                          Mismatch {mismatchSupplied = badHash, mismatchExpected = goodHash}
-                          (originalBytes <$> expectedIntegrity)
-                    ]
+                  submitFailingLegacySubTx lang badTopTx $
+                    SubScriptIntegrityHashMismatch @era
+                      Mismatch {mismatchSupplied = badHash, mismatchExpected = goodHash}
+                      (originalBytes <$> expectedIntegrity)
           it "the supplied hash is wrong" $ testHashMismatch . SJust =<< arbitrary
           it "the supplied hash is missing" $ testHashMismatch SNothing
 
         it "SubMalformedScriptWitnesses" $ do
           let scriptHash = hashPlutusScript $ asSLanguage slang malformedPlutus
           txIn <- produceScript scriptHash
-          submitFailingTx
+          submitFailingLegacySubTx
+            lang
             (mkTopTxWithSubTxs [scriptSpendingSubTx txIn])
-            [injectFailure . SubMalformedScriptWitnesses @era $ NES.singleton scriptHash]
+            (SubMalformedScriptWitnesses @era $ NES.singleton scriptHash)
 
         disableInConformanceIt "SubMalformedReferenceScripts" $ do
           script <- fromPlutusScript <$> mkPlutusScript (asSLanguage slang malformedPlutus)
@@ -380,6 +384,27 @@ spec = describe "SUBUTXOW" $ do
           submitFailingTx
             (mkTopTxWithSubTxs [subTx])
             [injectFailure $ SubInvalidMetadata @era]
+
+-- Legacy sub-transaction scripts also fail context translation, independently
+-- of the SUBUTXOW predicate under test. Fixup can change the sub-transaction id.
+submitFailingLegacySubTx ::
+  forall era.
+  DijkstraEraImp era =>
+  Language ->
+  Tx TopTx era ->
+  DijkstraSubUtxowPredFailure era ->
+  ImpTestM era ()
+submitFailingLegacySubTx lang tx expectedFailure =
+  submitFailingTxM tx $ \fixedUpTx ->
+    case OMap.elems $ fixedUpTx ^. bodyTxL . subTransactionsTxBodyL of
+      [subTx] ->
+        pure
+          [ injectFailure $
+              Conway.CollectErrors
+                [BadTranslation . inject $ UnsupportedScriptInSubTx @era lang (txIdTx subTx)]
+          , injectFailure expectedFailure
+          ]
+      _ -> assertFailure "Expected exactly one sub-transaction"
 
 -- | Every distinct reason a sub-transaction requires a key witness,
 -- paired with a sub-transaction that requires it and the key hash whose